Add the Digest Headers spec to SpecData.json
The Digest Headers spec isn’t in browser-specs, and may not be for a
long time yet (see https://github.com/w3c/browser-specs/issues/339) —
so this change adds it to SpecData.json.

Otherwise, without this change, it shows up as “Unknown specification” in
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Digest and
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Want-Digest

The change also drops the unnecessary arguments from the {{Specifications}}
macro calls in the above articles; the arguments aren’t needed because the
articles have browser-compat keys which the spec URLs can be picked up from.