Make Kitchensink table test case forward-compatible
This change makes the Markdown table test case in the Kitchensink page
forward-compatible. Otherwise, without this change, it’s using the
legacy/deprecated SpecName and Spec2 macros — and is a legacy “manual”
Specifications table (and we don’t have such tables in content any
longer; rather, the tables are generated based on the browser-compat or
spec-urls frontmatter metadata keys).