Move compat data into browser-compat key (in docs in API tree)
For the following reasons:

  - consistency with handling of BCD features in other existing docs
  - streamlining out unnecessary repetition of BCD feature names
  - in keeping with the principle of making use of frontmatter metadata
    to store data — BCD feature names - rather than having the BCD
    feature names in the prose body of documents

...this change takes BCD feature names out of the arguments to the
Compat and Specifications macros, and moves those BCD feature names into
the value of the browser-compat frontmatter metadata key.

Every update in this change follows the same simple pattern: It takes
the following:

> ## Specifications
>
> {{Specifications("api.SyncManager")}}
>
> ## Browser compatibility
>
> {{Compat("api.SyncManager")}}

...and changes that into this:

> browser-compat: api.SyncManager
> ...
> ## Specifications
>
> {{Specifications}}
>
> ## Browser compatibility
>
> {{Compat}}

...that is, it takes the `"api.SyncManager"` BCD feature name that’s
repeated in two different places in the prose body of the doc — as
arguments to the Compat and Specifications macros — and moves it being
given just once: in the browser-compat frontmatter metadata key.