Fix cases of Specification macro with unnecessary args
A number of Specifications macro calls have arguments that are
unnecessary because the spec URLs for the BCD feature in the argument
are already provided by the BCD feature given in the browser-compat key.

For example, files/en-us/web/css/align-content/index.md has this combo:

> browser-compat: css.properties.align-content
> ...
> {{Specifications("css.properties.align-content.grid_context")}}

Looking at https://github.com/mdn/browser-compat-data/blob/main/css/properties/align-content.json#L4
(that is, the BCD feature css.properties.align-content) in BCD, we find this:

> "spec_url": [
>   "https://drafts.csswg.org/css-align/#align-justify-content",
>   "https://drafts.csswg.org/css-flexbox/#align-content-property"
> ],

Looking at https://github.com/mdn/browser-compat-data/blob/main/css/properties/align-content.json#L472
(that is, the BCD feature css.properties.align-content.grid_context) in BCD, we find this:

> "spec_url": [
>   "https://drafts.csswg.org/css-align/#align-justify-content",
>   "https://drafts.csswg.org/css-flexbox/#align-content-property"
> ],

That is, the css.properties.align-content feature and its subfeature
css.properties.align-content.grid_context have exactly the same spec
URLs in BCD — so there’s therefore no need to give the Specifications
macro the css.properties.align-content.grid_context argument.

Specifically, removing that argument doesn’t cause any change in the
rendered output for the Specifications table in that document; with the
argument removed, the table still shows the same set of spec URLs.

All the updates in this change are for files where the source follows
the same pattern described above: the argument given to the
Specifications macro is unnecessary because it doesn’t change the output.