Replace some BCD tables with links
Two of the articles in this change have BCD tables whose features
the articles are not directly tied to — instead, the articles are
(inappropriately) using the mechanism just to show “here’s an example of
a BCD table”. So this change replaces the tables with inline links to BCD
tables in other existing articles (where users can view them in context).

The one other article in this change doesn’t really need to be showing
the BCD table it has, but can instead just link to the existing BCD table
that’s already in another existing article. So this change does that.