Move compat data in browser-compat key for input element attributes
For attributes that are specific to just the input element, this change
moves the BCD feature IDs out of the Compat macro call and into the
browser-compat key. The articles for these features have broken BCD
tables, so this change also updates/corrects the BCD features IDs to
match https://github.com/mdn/browser-compat-data/pull/16493. And that
BCD change makes it unnecessary for the MDN articles to have spec-urls
keys — we now can get those with the browser-compat key — so this change
also drops the (now-redundant) spec-urls keys.