Add a note on margin collapsing (#16745)
* Add a note about the margin collapsing

People may get the wrong impression that the margin collapsing is supposed to work on all sides of the element. Adding a note to clarify the intended behavior of the margin collapsing will come in handy.

* Update the note on margin collapsing

Rewrite the note on margin collapsing such that it focuses only on the positive side (i.e. vertical margins)

Co-authored-by: wbamberg <will@bootbonnet.ca>

* Rewrite the note on margin collapsing

* Rewrite paras on margin collapsing

Co-authored-by: wbamberg <will@bootbonnet.ca>