Django dev env: add short section on pylint-django
As discussed in https://github.com/mdn/django-locallibrary-tutorial/pull/105, I'm wondering if a small section to direct readers to `pylint-django` would be helpful for the tutorial. As of now a reader does `startproject`, likely gets files with errors, and may not understand that these errors are based on an incorrect linter being used.