Fix autoboxing example mutating a string
Running the example in the console does not print the expected output present in the comment (it should be `lower case string` instead of `LOWER CASE STRING`). This behavior is expected since primitive types are immutable.

However since the goal of this section is to show that it's possible to automatically call a function due to autoboxing, I think it's better to keep the `LOWER CASE STRING` and I fixed the code to make it produce the desired output. I am not sure if it would be cleaner to put the result of `toLowerCase()` in a variable or not, but I wanted to keep the example simple.

(This issue seemed related to #14626 but it is not)