Make KeyboardEvent.Code page keyboard accessible
Chromium browsers don't make iframes focusable by default.
This means that keyboard-only users on Chromium browsers cannot
interact with the examples on this page unless they have a tabindex.
This shouldn't be a Chromium issue because the HTML specifications
don't have a normative "default tab index" for iframes as far
as I'm aware.