Drop (for now) spec URL and BCD link for elementtiming HTML attribute
There is no feature for the elementtiming HTML content attribute in BCD
yet — not at html.elements.attribute.elementtiming no anywhere else in
BCD — and it is as yet not actually defined in any spec; the spec URL
https://wicg.github.io/element-timing/forms.html#attr-label-for is wrong
— that spec has no forms.html page, and that attr-label-for anchor would
anyway not be the right anchor for the feature.

https://github.com/whatwg/html/issues/8024 is a related issue for
creating a normative spec definition for the feature. After the feature
is actually defined in a spec and added then added to BCD, we could
include the Specifications and Compat sections again.