Updating variable names that use "json" nomenclature
NOTE: This commit is probably a conversation-starter.  If you agree with the spirit of the changes, then other variable names in this file might also need to be update for consistency.  

ISSUES BEING ADDRESSED:  
(1) response.json() returns a regular JS object (not json).  See: https://developer.mozilla.org/en-US/docs/Web/API/Response/json.  So for new learners (and old learners like me!), seeing a variable named "json" is quite misleading if it is not actually json.  For now, I just changed a few lines in the Async/Await section that used the variable name "json" (but we may also want to update other sections of this file for consistency reasons).

(2) The last example under Async/Await used a ".then()" in an example which is focused on using async/await.  For clarity to new learners, I changed it to use await.