Drop “Difference between log() and dir()” section from console.log doc
Fixes https://github.com/mdn/content/issues/17444

As noted in https://github.com/mdn/content/issues/17444, the content of
this section was copied verbatim from a Stack Overflow answer. The
content is also specific to Chrome. So it’s best to just remove the
entire section — there’s a lot of value to it.