Fix width measurement (#17886)
> positive numbers indicating a distance going left from the given alignment point

Therefore, we should directly add `textMetrics.actualBoundingBoxRight` and `textMetrics.actualBoundingBoxLeft`. When the text is partly out of the left boundary, textMetrics.actualBoundingBoxLeft is positive. When the text doesn't touch the left boundary, the value is negative. 

When the text doesn't touch the left boundary, the original algorithm will report wrong answer. (Although I don't know if there will be a situation like this, but I think making it always right is better.)