Update index.md (#18180)
'small' is a subjective description. I would suggest that the course fees (showing as 74 EUR) are not 'small', so propose to omit this description. Furthermore the 'one-off' payment description is not completely accurate as access is provided for 5 years. Whilst this is pretty long, it's not truly a one-off payment, the payment is for 5 years of access. Propose to remove both of these statements, but retain the link to a valuable resource.