update way of explaining how to add methods to prototypes
The old code overwrites the entire prototype with an object deleting the old constructor, and then adds the constructor on another line. I, who was studying the subject found it quite complex, and there are simpler ways to do it without having to remove and add a constructor again, and explain superficially what is going on.

It could be even more simply:
```js
function greet() {
  console.log(`hello, my name is ${this.name}!`);
}

function Person(name) {
  this.name = name;
}

Object.assign(Person.prototype, {greet});
// or
//Person.prototype.greet = greet;
```

But then it would be difficult to link with the previous example and would change the texts a lot. Not to mention that it might have to quote destructuring to better understand what is happening.

It's a suggestion to simplify and not have to address unnecessary things to the topic explained in the snippet. It's much easier to understand an Object.assign than to understand why you need two lines because of an overriding that was done.