remove outdated hgroup content (#18530)
The [`<hgroup>` element](https://html.spec.whatwg.org/multipage/sections.html#the-hgroup-element) was recently significantly revised in the HTML specification. The warnings about its impact on the outline algorithm are no longer relevant, as the concept of the outline algorithm was removed from the HTML specification as well.

This PR simply removes outdated content, and replaces the remaining bits with content that is more inline with the current definition for the element.