Remove reference to linebreaking attributes (#18576)
Documentation for <mo> and <mstyle> mention linebreaking attributes from
MathML3 but they have never been implemented in any browser and they are
currently not part of MathML Core. It's not clear if in the future
linebreaking would actually follow the same design as MathML3, so let's
remove these for now to avoid confusing web developers.