Remove Firefox-specific notes for the <math> element
The first note essentially says that some <mstyle> attributes were
implemented on the <math> element in Firefox 7 and 8. Nowadays, these
are actually global attributes so they are no longer mentioned on the
specific article for the <math> tag. Moreover [1] currently says

"Prior to Firefox 70, the attribute was only accepted on a few elements,
as specified in MathML 3."

so that seems enough information, no need to be more specific about
when the attributes were implemented on the <math> tag.

The second note mentions alttext/altimg/altimg-width/altimg-height,
altimg-valign from MathML3. These are actually not part of MathML
Core, not implemented in browsers (they would instead expose an a11y
tree as described in [2]) and not even mentioned on the article itself.
We can add this back if it turns out that these are added in future
MathML Core versions, but for now let's remove them to avoid any
confusion for web developers.

[1] https://developer.mozilla.org/en-US/docs/Web/MathML/Global_attributes
[2] https://w3c.github.io/mathml-aam/