Update index.md
I believe an idempotent function *can* have side-effects as long as those side effects are also idempotent. This document suggests otherwise.  POST is not idempotent since by definition is is changing the database state every time you call it, but PUT can be as long as repeatedly calling that PUT with the same arguments results in the same database state.  If you are updating a counter of updates for a given record even if the values are the same, only then would that PUT not be idempotent.