Improve description Result section
It was not immediately obvious how the resulting page works, so the added description describes the behavior more fully.

I think it would also be a good idea to shorten the length of time an ad stays on screen to, say, 10 seconds, for this demonstration. However, I don't think the article needs to be changed, a minute is more reasonable for the use case, however, I assume that production code would have variable length ads.

I am not sure where the Building_the_site code is so I cannot make a PR on that.