Fix a typo
When an inherited property is changed down in the prototype chain we say that the proprty is shadowed. Here, in that context, all the objects, by default, inherit hasOwnProperty which is a function, so we have to say : 
All objects, except those with `null` as `[[Prototype]]`, inherit hasOwnProperty from `Object.prototype` unless it has been overridden further down the prototype chain.