Clarify how to format a datetime-local value
The existing language suggests that one can take a `Date` object, call `toISOString()` on it, and use the resulting string as the value for a datetime-local input. This doesn't work, because `toISOString()` includes a "`Z`" at the end to indicate the timezone, while datetime-local expects the value to have no time-zone offset information. See https://html.spec.whatwg.org/dev/input.html#local-date-and-time-state-(type=datetime-local).