Add hint how to experiment with TRY IT box
I came to MDN to this page expecting to find some information about `vh`. However, it is not mentioned here, and I think it should probably be mentioned here with a link to https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_Values_and_Units , for example.

If this is a good idea, then I would be happy to review this page to see what else needs to be updated and make another PR if necessary.