Removed references to sectioning roots, <menu> types & fixed <input> anchor link
The page contains references to sectioning roots and `<menu>` type attributes, but both seem to have been depreciated. Also, a link to the `<input>` page used `htmlattrxref`, so its anchor was broken. I removed the references to sectioning roots and `<menu>` type attributes and  replaced all the anchor for the `<input>` page link.