Improve the fallback example for :focus-visible (#19993)
* Improve the fallback example for :focus-visible

After replacing the custom button with a native button in the previous change to this example, the styles of the example now affect the native styles of the button. It is better to move the properties that change the appearance of the button itself into its general styles.

In addition, this example has always been difficult to understand for beginners due to different styles for different focus pseudo-classes. Making the example look more realistic and practical will make it more useful for developers.

* Update the section heading

* Improve text

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

* Improve code comment

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

* Improve fallback comment

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

* Remove unwanted part

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>

* Fix test failure (sorry)

Co-authored-by: Jean-Yves Perrier <jypenator@gmail.com>