Update comment for string.replace() code example
The example on line 73 has no explanation for the output, whereas the examples surrounding it did. 
The behaviour shown in line 73 is explained in the text above, but that is also the case for the other examples . For consistency I have added 
`the pattern is a string, so it doesn't have any groups` as an explanation.