Change one of the extensions mentioned (#20163)
The article suggests overriding styles being easier through the use of extensions, before linking to Stylish for Firefox and Chrome. According to users of the Firefox version, the extension has been abandoned by it's original creator and has been turned into spyware. Whether this is the case or not, I felt linking to Stylus, a recommended  extension, would be a better fit.