Fix legacy algorithm for <semantics> and remove obsolete attributes
Update the note about the legacy algorithm implemented in WebKit and
Firefox, which actually skip the annotations that have an `src`
attribute [1] [2]. Also try to clarify that <annotation-xml> without
a known encoding are skipped too, as current text is not clear.

MathML Core only mentions `encoding` for non-global attributes
accepted on annotation/annotation-xml elements. So mark `src` as
deprecated and remove all the other attributes that were never
involved in browser implementations.

[1] https://github.com/WebKit/WebKit/blob/b7d555805988f56bc97e9ae21014a35a463b1987/Source/WebCore/mathml/MathMLSelectElement.cpp#L174
[2] https://searchfox.org/mozilla-central/rev/3f9dcc016dd96a0336d46f4a19aeabdd796ab9e9/layout/mathml/nsMathMLsemanticsFrame.cpp#63
[3] https://w3c.github.io/mathml-core/#semantics-and-presentation