Update links from RFC 7540 to RFC 9113 (#20740)
There were some specific links to RFC 7540, these can be updated.

At the same time, the notes mention a topic that applies equally to HTTP/3. So add that too. However, the specific comment about HTTP/2 compat is not expanded to cover HTTP/3, since there is not evidence that I am aware of.