rewrite confusing statement
The original statement is difficult to understand: "Having modified network requests, wide open to man in the middle attacks would be really bad."
Modified network requests _are_ a man in the middle attack. The point missing here is that HTTP allows for a man in the middle attack which modifies the network request to inject a service worker, and that would be really bad given how powerful these APIs are.