Update index.md
In the docs on Inheritance, https://developer.mozilla.org/en-US/docs/Web/JavaScript/Inheritance_and_the_prototype_chain,
it is recommended not to reassign the prototype of a constructor function because it points the prototype to a new object and old instances will not be able to access the newly assigned prototype. It is recommended to use setPrototypeOf.