Fix a bunch of typos in what_went_wrong
Page says to look for `const lowOrHi = document.querySelector('lowOrHi');` at line 48. It doesn't correspond to where this line is actually placed. It is placed on line 49. This change fixes this and makes the instructions consistent with the code in [number-game-errors.html](https://github.com/mdn/learning-area/blob/main/javascript/introduction-to-js-1/troubleshooting/number-game-errors.html)