Confusing Terminology
Rejected promises are `rejected` and not `resolved`. Using the term `"resolved"`(which is typically used for resolved promises) interchangeably for rejected promises is confusing, misleading, and inconsistent with the terminology used to describe the settled state of promise objects.

The below excerpt mentions that a `rejected` promise is `resolved` to the value `88`. 

 - Similarly, unlike `Promise.reject(3).then(() => {}, () => 88)`
    (which will return a resolved promise with the value `88`),
    `Promise.reject(3).finally(() => 88)` will return a rejected promise
    with the reason `3`.