Wording should unambiguously match implementation (#21232)
* Suggest wording change to match implementation

Based on the ECMAScript Language Specification description of Array.findLastIndex, Array.findRight and Array.findIndexRight would be more accurate names.

I tried to use Array.findLast to find the value of the last element in an array that passed a test whose condition could change during iteration. Instead, it found the value of the first element that passed the test, going from the end of the array to the beginning of the array. The name "findLast" is ambiguous, since "last" could mean the last item that passed a test, where all items were tested, or it could mean the item closest to the end that passed a test.

* Update Array.findLastIndex

* Use review suggestion

* Update TypedArray.findLast

* Update TypedArray.findLastIndex

* Remove inconsistent term

* Remove inconsistent term