Update index.md
https://github.com/mdn/content/pull/21368#issuecomment-1271868459

Hi @wbamberg Thanks for reviewing my PR. I can change the const to a let and then it will be pastable. I think it's made clear that a promise is returned within the previous section where the error message is shown ``` "promise" is a Promise object ``` but what is interesting is that the same section tries to highlight how async/await can be used to work "just as if ```fetch()``` were a synchronous function!" but then the examples make it seem that it's only true up to a certain point, which is not the case. I think it's important that the reader understands that there is a valid way of writing the example in the synchronous way it was trying to.