Correct `role="alert"` information
The information currently on https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/alert_role is, unfortunately, outright wrong. Like all live regions, the container with `role="alert"` needs to first be "primed". After that, any *subsequent* content updates are announced. It doesn't work (or, when it does, it's mostly a fluke/down to internal browser timings) to just dynamically show a container by flipping its `display` property, or adding `role="alert"` to an existing element, or having that `role="alert"` and the message already present in the static markup on page load.

This PR corrects both the prose, and replaces the broken/incorrect examples with working/correct ones.