fix: Remove unrelated code blocks
This is definitely opinion-based, but I think these code blocks
represent a non-sequitur. They seem to be just pointing out that you
can store a reference to a DOM node before adding an event listener.
To me that is more about how variables work and seems to only
tangenially concern event handlers.