Cross-reference destructuring assignment
I knew that destructuring assignment existed, and that it looked like an extended form of assignment, but I didn't remember what it was called.  In my mind it's not a different *operator*, but rather a more advanced form of *this* operator, so link to a list of assignment operators didn't seem like what I wanted.