Update Transfer-Encoding to hex numbers
The current chunked encoded example is streaming strings that their length is lower than 10 (Mozilla, Developer, Network), therefore, it is easy to miss that the chunked length should be in hexadecimal form.

By merging the 2nd and 3rd line into `Developer Network`, we reach length of 17 which is 11. This marks it much clearer that it is hex.