Remove duplicate warning about -webkit-box-reflect (#21986)
* Remove duplicate warning about -webkit-box-reflect

This page features the common “non-standard” header, as well as a section under “Specifications” that covers the same ground. I don't believe a second warning is necessary.

I also think the warning itself is also potentially misleading:

> This feature is **not intended to be used by Web sites**

According to who? This is a very subjective opinion, phrased as an objective statement.

> To achieve reflection on the Web, the standard way is to use the CSS element() function.

I suppose this is true, if we're defining “standard” as “in the CSS specification”. But, right now, `element()` is only supported by Firefox ([~3.5% of global traffic](https://caniuse.com/css-element-function)), while `-webkit-box-reflect` is supported by every other major browser ([94.5% of global traffic](https://caniuse.com/?search=-webkit-box-reflect)). And so, the standard way that developers are _actually implementing_ reflections is with `-webkit-box-reflect`, and it's pretty much the only option if you want _most_ people to see it, but you'd never know it from this warning box.

I understand the desire to push people towards the standards-aligned way of doing things, but I think that only helps when the standards-aligned way is well-supported by browsers.

* Remove warning