DOM whitespace: show 0th step of example processing
The examples are shown a bit earlier in the document,
sufficiently so that they can be above the fold
when one gets to the sequence of steps.

Furthermore, the sequences of steps
include the visual whitespace indicators directly in the HTML code,
unlike what's done in the initial example code blocks.

For pedagogical purposes, it's helpful to address both these issues,
by showing the example as a zeroth step,
so that one can see the source text be gradually transformed.

Also, fix a small typo: "the may be" → "there may be".