Improve substring description (#22559)
The previous version of the text made it reasonable to assume that if the text would be "Mozilla" and the indexes `(1, 3)`, the end result will be the character `z`,
i.e. the string "between the start and end indexes".

Since that would be very unusual, many users might instead assume that it would result in `ozi`. But that would also be incorrect.

This change makes it clear that the function includes the first index and excludes the second.