Update index.md
I thought this added description that I am proposing illustrates what we are instructing JavaScript to do via our code if indeed it is not the user's first attempt at guessing the number. My proposal sheds light on how we are using  template literals to pass in information while adding it to itself. I believe this will come in handy for new learners to keep in mind. Especially when thinking in terms of being able to use our variables in a more creative and advanced manner in coming lessons. If I am misinterpreting how we are using JavaScript in this context, would you please notify me and point my attention to what I am getting wrong in terms of reassigning and using template literals? Thank you kindly. 
Joseph
teacherjosephg@gmail.com