Clipboard.write: document file type limitations
Clipboard.write is restricted to text, HTML, and PNG data, with this limit enforced here:
https://searchfox.org/mozilla-central/source/dom/events/Clipboard.cpp#523

That function is called from here:
https://searchfox.org/mozilla-central/source/dom/events/Clipboard.cpp#536

And that is in turn called by the core Clipboard.write function, here:
https://searchfox.org/mozilla-central/source/dom/events/Clipboard.cpp#629

This limit really should be documented up-front. The API does have an error message for this, but it doesn't list which formats are accepted; worse, this is a limit you can easily run into when developing WebExtensions, and JS error reporting for WebExtensions has been broken for the last five years (see Bugzilla bug 1410932).