Update WebGL1 example
Hi, I was using this documentation for reference and found out that the overload `texImage2D(target, level, internalformat, format, type)` is actually not supported by either Firefox or Chrome. It is not even part of the WebGL1 specification: https://registry.khronos.org/webgl/specs/latest/1.0/#5.14.8.

Also, `texImage2D(target, level, internalformat, format, type, pixels)` does not work with TypedArray or a DataView or null  so just for clarity I propose also updating the comment in the example.