autofocus a11y updates
follow on to the merged #23009

minor wording adjustments and some clarifications to the a11y-related content for `autofocus`.  
Removed the last sentence about 'tabbing order should always follow the dom order' as that's not 100% always true, nor is it applicable here (tab order _does_ follow DOM order, even with `autofocus` in use - it's just the element that's initially focused in that order will be adjusted per use of the attribute).