Address Resize Observer error events
This revision provides the rationale for the error, as well as potential consequences. Hoping to reduce time to root cause & determine course of action for authors seeing this error, as I've seen many bug reports mentioning this error.
Also suggests the upcoming container queries as a potential alternative that does not expose authors to the infinite loop problem.