Move “Transferable objects” out of Glossary and into Web/API tree
See https://github.com/mdn/content/issues/23099#issuecomment-1359834232.
If this content were to remain the Glossary, that would continue to
prevent us from being able to refine it further — because, as noted,
glossary entries are meant to be brief, and this content is already much
too long to be appropriate for having in the Glossary.