Fix comment in a code example
The comment in JWK code example was about PKCS #8 import above. It's changed to refer to JWK, similar to https://github.com/mdn/dom-examples/blob/main/web-crypto/import-key/jwk.js
In the link above, the function is said to take a string, when in fact it takes a JsonWebKey object: "a string" was corrected to "an object representing".