CSS `counters()` - remove note about type-or-unit
This note appears to be inaccurate. There is no type-or-unit parameter in the spec, and support for `counters()` in places other than `content` does not appear to be experimental. (There's even a statement earlier in the article that `counters()` "can be used, theoretically, anywhere a `<string>` value is supported"). There is no corresponding note in the browser compatibility table.

See also https://github.com/mdn/content/pull/12562, which fixed this same error in the corresponding article on [`counter()`](https://developer.mozilla.org/en-US/docs/Web/CSS/counter).