Use accurate language for BigInt.asIntN()|BigInt.asUintN()
5f174593bf1b8b97 already updated this language to use “clamp” instead
of “wrap”. However, this was a misleading change, since “clamp” is
generally _always_ used to indicate an entirely different operation
(namely, `Min(Max(·, a), b)` for some `a` and `b`) rather than what
these functions actually do (namely, `· mod 2^n`).

Some examples of how “clamp” is used to refer to this other
operation include:

- https://developer.mozilla.org/en-US/docs/Web/CSS/clamp
- https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Uint8ClampedArray
- https://en.wikipedia.org/wiki/Clamping_(graphics)
(and, of course, the actual “clamp” tool that gave rise to this metaphor
in the first place.)

I agree that “wrap” was maybe not the most intuitive choice of word
here. Consequently, this commit uses “truncate”, a more accurate and
hopefully also a more intuitive word, and rewords the introductory
sentences and examples to be more precise instead of using a choice of
word that JS developers may not be familiar with anyway.