Formatting box → Content box
I could be completely wrong on this one.

I have not come across the term "formatting box" before. My understanding is the term "formatting box" is being used here to refer to the standard content (what would usually go in the "content box") with the addition of `::before` and `::after` elements. Given that I could not find the term "formatting box" in the spec, I thought it might be clearest to use the more common box model term "content box" as the `::before` and `::after` elements are rendered inside of an elements content box. It then felt best to update the link people to point people to the standard box model page